requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
typing-extensions==4.5.0
python-multipart==0.0.6
pandas==1.5.3
gunicorn==20.1.0
numpy==1.23.5
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
requests==2.31.0
langchain==0.0.211